+++ src/Corvus.Text.Json.CodeGeneration/CodeFileBuilders/CorePartial.cs
@@ -49,7 +49,6 @@ public sealed class CorePartial : ICodeFileBuilder
                 .BeginTypeDeclarationNesting(typeDeclaration)
                     .AppendDocumentation(typeDeclaration)
                     .AppendDebuggerDisplay()
-                    .AppendUnionAttribute(typeDeclaration)
                     .BeginPartialStructDeclaration(
                         typeDeclaration.DotnetAccessibility(),
                         typeDeclaration.DotnetTypeName(),
@@ -58,7 +57,6 @@ public sealed class CorePartial : ICodeFileBuilder
                             "IFormattable",
                             new("ISpanFormattable", FrameworkType.Net80OrGreater),
                             new("IUtf8SpanFormattable", FrameworkType.Net80OrGreater),
-                            .. typeDeclaration.UnionMembersInterfaces(),
                             ])
                         .PushJsonSchemaClassNameAndScope()
                         .PushJsonPropertyNamesEscapedClassNameAndScope()
@@ -111,7 +109,6 @@ public sealed class CorePartial : ICodeFileBuilder
                         .AppendFreezeMethod(typeDeclaration)
                         .AppendTryGetAsCompositionTypeMethods(typeDeclaration)
                         .AppendMatchMethods(typeDeclaration)
-                        .AppendUnionMembers(typeDeclaration)
                         .AppendJsonPropertyNames(typeDeclaration)
                         .AppendJsonPropertyNamesEscaped(typeDeclaration)
                         .AppendJsonPropertyNamesPrebaked(typeDeclaration)
@@ -124,6 +121,7 @@ public sealed class CorePartial : ICodeFileBuilder
                         .PopJsonPropertyNamesPrebakedClassNameAndScope()
                         .PopJsonSchemaClassNameAndScope()
                     .EndClassStructOrEnumDeclaration()
+                    .AppendUnionPartial(typeDeclaration)
                 .EndTypeDeclarationNesting(typeDeclaration)
                 .EndNamespace()
             .EndFile(typeDeclaration, string.Empty);